requirements.txt
dnspython==1.16.0
google-cloud-storage==1.23.0
google-cloud-pubsub==1.2.0
psycopg2==2.8.4
pydantic==0.32.2
requests==2.22.0
responses==0.10.7
python-dateutil==2.8.1
pandas==1.0.1
